+++ programs/drift/src/state/perp_market.rs
@@ -22,7 +22,7 @@ use crate::math::safe_math::SafeMath;
 use crate::math::stats;
 
 use crate::state::oracle::{HistoricalOracleData, OracleSource};
-use crate::state::spot_market::{SpotBalance, SpotBalanceType};
+use crate::state::spot_market::{AssetTier, SpotBalance, SpotBalanceType};
 use crate::state::traits::{MarketIndexOffset, Size};
 use crate::{AMM_TO_QUOTE_PRECISION_RATIO, PRICE_PRECISION};
 use borsh::{BorshDeserialize, BorshSerialize};
@@ -58,7 +58,7 @@ impl Default for ContractType {
     }
 }
 
-#[derive(Clone, Copy, BorshSerialize, BorshDeserialize, PartialEq, Debug, Eq)]
+#[derive(Clone, Copy, BorshSerialize, BorshDeserialize, PartialEq, Debug, Eq, PartialOrd, Ord)]
 pub enum ContractTier {
     A,           // max insurance capped at A level
     B,           // max insurance capped at B level
@@ -67,9 +67,26 @@ pub enum ContractTier {
     Isolated,    // no insurance, only single position allowed
 }
 
-impl Default for ContractTier {
-    fn default() -> Self {
+impl ContractTier {
+    pub fn default() -> Self {
         ContractTier::Speculative
+    }
+
+    pub fn is_as_safe_as(&self, best_contract: &ContractTier, best_asset: &AssetTier) -> bool {
+        self.is_as_safe_as_contract(best_contract) && self.is_as_safe_as_asset(best_asset)
+    }
+
+    pub fn is_as_safe_as_contract(&self, other: &ContractTier) -> bool {
+        // Contract Tier A safest
+        self <= other
+    }
+    pub fn is_as_safe_as_asset(&self, other: &AssetTier) -> bool {
+        // allow Contract Tier A,B,C to rank above Assets below Collateral status
+        if other == &AssetTier::Unlisted {
+            true
+        } else {
+            other >= &AssetTier::Cross && self <= &ContractTier::C
+        }
     }
 }
 
@@ -550,7 +567,8 @@ impl AMM {
     pub fn get_oracle_twap(&self, price_oracle: &AccountInfo) -> DriftResult<Option<i64>> {
         match self.oracle_source {
             OracleSource::Pyth => Ok(Some(self.get_pyth_twap(price_oracle, 1)?)),
-            OracleSource::Pyth1000 => Ok(Some(self.get_pyth_twap(price_oracle, 1000)?)),
+            OracleSource::Pyth1K => Ok(Some(self.get_pyth_twap(price_oracle, 1000)?)),
+            OracleSource::Pyth1M => Ok(Some(self.get_pyth_twap(price_oracle, 1000000)?)),
             OracleSource::Switchboard => Ok(None),
             OracleSource::QuoteAsset => {
                 msg!("Can't get oracle twap for quote asset");